EnsureBrowserContextKeyedServicesFactoriesBuilt -- Chrome common Factories
Add missing common Chrome KeyedServiceFactories to the initialisation
list in `EnsureBrowserContextKeyedServicesFactoriesBuilt()` to properly
populate the dependency graph at start.

If the factory `GetInstance()` method does not exist, is private, or
the whole factory is private (or implemented in a .cc file), only expose an `EnsureFactoryBuilt()` method in the relative service or
code that matters.

Bug: 1412439
Change-Id: I643473899914b9c28886e9511138881265b243b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4190213
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1102058}