[Shortcuts] Deprecate shortcuts in App struct from crosapi.
This CL deprecates the shortcuts array that live in App struct from
crosapi. This is in preparation to remove shortcuts struct out from the
App struct, because we are planning to have a separate cache to store
shortcuts. (see go/shortcuts-system-dd).

For the mojom struct, the shortcut struct mapping is also removed
because it will be hard to change the C++ struct otherwise. This change
is save because it end up giving an empty array in the apps::App struct
and this field is not used in the code anyway.

BUG=1412708

Change-Id: I575f09fee82497b48b1d9a96e5bfa9871239a0dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4221102
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Auto-Submit: Maggie Cai <mxcai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102064}