[Signin] Resize Signin Intercept bubble dynamically based on content
This CL implements a way of resizing DiceWebSigninInterceptionBubbleView
based on the content of the page.

The widget is not shown anymore until the view is loaded. After it is
loaded, the final height of the page is passed to
DiceWebSigninInterceptHandler, which runs a callback to set this height
to the view and to the widget, and the widget is then shown.

Screenshots: https://crbug.com/1412110#c2
Demo:
https://screencast.googleplex.com/cast/NjMzMzcwODE1NzA1OTA3Mnw1NGFiYTViMC04Yg

Bug: 1412110
Change-Id: I22d8e5005df9c162093dacb9f023f8463cf2ff01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4210262
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Gabriel Oliveira <gabolvr@google.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1102073}