[views] Adding Fab to views_examples
This is a prototype of how a Floating Action Button could potentially
be implemented in the Views library.

Proposal is to shrink the size of the button but the margins of the shadow
and draw a shadow whenever the mouse hovers over the button.

http://screencast/cast/NTAyNDA4NDMxMzI0MzY0OHxkNjhhY2IyZi02ZQ

Change-Id: Icef39e51edace0d1abfd0b223b2d8bf2a4ba22df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4209940
Commit-Queue: Allen Bauer <kylixrd@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102142}