desktop-screenshots: remove editor support
We decided, after a great deal of UX feedback, that this component isn't
polished enough to ship and we don't quite see a path to getting it
there without a big rework. This change deletes the feature, the logic
to install the component, and the wiring for it in the screenshot
bubble.

Bug: 1410277
Change-Id: I7e47ece043e3fe1f374c3c9435e3cdb54b9b0131
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4198648
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Chris Bookholt <bookholt@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Sophey Dong <sophey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102193}