Add VRR helpers to DisplaySnapshot
This CL adds convenience accessors to DisplaySnapshot for performing
common checks about whether a display is vrr-capable and/or vrr-enabled,
which are otherwise cumbersome and verbose to perform.

This CL additionally removes the const modifier from the VRR state
property of DisplaySnapshot and adds a corresponding setter so the it
can be updated after a successful configuration.

Bug: 221220344
Change-Id: Ib9250dd32b0da0fdb8ed4d12e9d0c887b733d7fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4220194
Commit-Queue: Daniel Nicoara <dnicoara@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Auto-Submit: Andrew Wolfers <aswolfers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102195}