DanglingPtr: triage UnretainedDanglingPtr in GrpcHttpConnectionClient
Doc:
https://docs.google.com/document/d/1YQCh_mezoSF1fWS9TaMEh0Whumb8eNe0KWV8MBs7sFs

We detected those callbacks are invoked using dangling pointers.
This is something we would like to detect and potentially ban.

In the meantime, we need to annotate them using:
base::UnsafeDanglingUntriaged

Outside of the experiment, it is a no-op.
Within the experiment, it skips DumpWithoutCrashing.

Bug: 1291138,1413634,1413547
Change-Id: I2a62df92d9ca038835b737668e3f7212a4632496
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4227732
Commit-Queue: Paul Semel <paulsemel@chromium.org>
Reviewed-by: Tao Wu <wutao@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102246}