Fenced frame: send automatic beacon for top navigations.
This CL is a follow up to the following CL:
https://chromium-review.googlesource.com/c/chromium/src/+/4167776
which adds `fence.setReportEventDataForAutomaticBeacons`.

When a top navigation occurs that's initiated by a fenced frame (defined
by a fenced frame navigating a frame outside of the fenced frame tree),
a beacon with the event type `reserved.top_navigation` will be
automatically sent out. It will be populated with the event data and
sent out to the destinations that were specified using
`fence.setReportEventDataForAutomaticBeacons`.

The beacon is sent out right after the navigation begins and the
NavigationRequest object is created. This will allow for a follow up CL
to send information about the NavigationRequest to ARA for processing.

Change-Id: I575b8c7d78c95ed478ef931fe3baa1624367c86c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4120753
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Liam Brady <lbrady@google.com>
Cr-Commit-Position: refs/heads/main@{#1102251}