[GridNG] Clamp auto-placement to subgrid area
While prototyping auto-repeat for subgrids, I came across a subtle
placement bug with subgrids that's a sub-case of
repeat-auto-fill-001.html.

In summary, auto placement for grids depends on `minor_max_end_line_`,
which determines when to wrap lines. For normal grids this can extend
indefinitely, and implicit tracks will be created to fit the expanded track numbers. However, for subgrids, this can be at maximum the
subgrid span size. Implicit tracks can be created for subgrids, but
only up to the subgrid span size.

Prior to this change, we were only clamping at the very end of
placement, which was causing various "double-placement" issues,
demonstrated in the test case added.

Bug: 618969
Change-Id: Ic3972408e4dd4f2069fa1bf26c95d815c26cb929
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4225602
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1102306}