Handle cancel result from start of printing
When a print job is started there can sometimes still be some user
interaction.  One such case can happen on Windows when the printer
driver is configured to print to port FILE:.  Should a user choose
to cancel at that point instead of selecting a file to print to,
then the print job should terminate, but should not display an
error message.

Update the lowest levels of Windows printing context code to handle
the ERROR_CANCELLED error code.  Modify the PrintJobWorker and PrintJob
to handle canceling in this scenario, and make this status changes
available to observers.  PrintViewManagerBase can then use this to be
aware of the canceling state and thus not show an error dialog.

Low-Coverage-Reason: Windows tests restricted by crbug.com/1256506
Bug: 1393505
Change-Id: I9acecb77c3868ca6c284919de9d7a0197c2cd9ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4220482
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Reviewed-by: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102322}