devtools: lookup for record definitions to avoid handling declaration
we are not interested by declarations since we want to lookup record
fields. Furthermore, a recent behaviour change in llvm
(https://reviews.llvm.org/D142384) makes the `fields()` method returning
fields even when the record is a declaration, which was not the case
before.

Bug: 1412769
Change-Id: Ibee339e067a73567bfaf12affbb4050afd37e07a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4225652
Reviewed-by: Hans Wennborg <hans@chromium.org>
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Commit-Queue: Hans Wennborg <hans@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1102335}