[ZBB][SPC] Align ClientDataJSON topOrigin output across platforms
In the Android implementation generating the clientDataJSON output
string for SPC, the topOrigin field has a trailing slash, whereas the
desktop implementation does not.

On desktop, a Serialize method[1] is used, which does not add a trailing
slash. Android does not have an equivalent method (there is
GURL::serialize, but this adds serializer information meant to be used
with GURL::deserialize), so GURL::getSpec is used, which adds the
trailing slash. This is fixed in this patch by sending the Origin over
JNI and formatting the string using the same C++ method. This is now
tested by adding a unit test for ClientDataJson, which was previously
untested as it is mocked in Fido2CredentialRequestTest.

This patch also aligns the origin source used on android with desktop;
i.e. the origin of the main frame, which may be opaque in a sandboxed
frame. This is tested by verifying the mocked call to
buildClientDataJson in Fido2CredentialRequestTest.

[1]
https://source.chromium.org/chromium/chromium/src/+/main:content/browser/webauth/authenticator_common_impl.cc;l=961-964;drc=c14f4e9399fc4e23a6c4c68fc60086bdceec1c90

Bug: 1394756
Change-Id: I3bbc5210185bf1bf543de74dfd7aafefaf395811
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4197639
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Nick Burris <nburris@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102343}