Keep kCompatibleVersionNumber in sync with kCurrentVersionNumber
The compatible version number is checked to see if the db file can be
used at all with the current code. So if the version is updated with
a backward-incompatible schema change, the compatible version should
also be updated to avoid the case if user reverts to an old version
of code and operates on the incompatible schema and crash.

Also bump up the schema version to 5 to make sure the databases with
the problematic `compatible_version` can be updated correctly during
migration.

Bug: 1409117, 1413787
Change-Id: I740c4be7e3216f1a8dbed593febd88bde705c923
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4225211
Reviewed-by: Chris Fredrickson <cfredric@chromium.org>
Commit-Queue: Shuran Huang <shuuran@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1102344}