Fixed the issue with the inactive dual internal mic
This CL fixed the issue that the inactive dual internal mic doesn't show
up in the audio detailed view. This is caused by the fact that we don't
save the volume level for inactive dual internal mic. We decided to use
the front mic's level as the default value for it if it's inactive. This
will not cause any issue since if the dual internal mic becomes active,
it will automatically show the actual value. A unit test is also added
for this situation.

Screenshot:
https://screenshot.googleplex.com/At7ycqiuXAuYTZA

Bug: b/267355114
Change-Id: I9bd7e7ebee245344eb5d733224fa123809bfeed6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4222467
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Commit-Queue: Sylvie Liu <sylvieliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102363}