Multiplanar: Move RasterDecoder::CopySharedImage to CopySIHelper
This is reland of [1]. The original change caused some failures [2] on
Android bot due to missing DrDC usage account in SubmitIfNecessary.
This change is a follow-up on the change [3] that accounts for
if DrDC is enabled when submitting commands to Skia.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4204147
[2] https://bugs.chromium.org/p/chromium/issues/detail?id=1412168
[3] https://chromium-review.googlesource.com/c/chromium/src/+/4226692

Bug: 1366486
Change-Id: I7843501f4c4e7b1850978bdf73a800cf32ff5b88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4219966
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1102374}