Reland "iwa: Remove WebApp::is_storage_isolated_"
This is a reland of commit a3abed430fa5cb26f4dc10414f26e4c116265015

The original version of this CL updated
WebAppFrameToolbarBrowserTest_Borderless to mess with WebAppProvider
directly by casting it to a FakeWebAppProvider, but in browser tests
we use the real WebAppProvider so the cast was incorrect. This was
caught by a UBSan builder. To fix it this CL uses ScopedRegistryUpdate,
but eventually the test should be rewritten to call the IWA
installation command instead.

Original change's description:
> iwa: Remove WebApp::is_storage_isolated_
>
> This field reflected the isolated_storage manifest field, which was
> removed in crrev.com/c/4214426.
>
> Bug: 1412210
> Change-Id: I60aedec02fe4bb1349cdb675a12003911a3d53e2
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4215562
> Reviewed-by: Phillis Tang <phillis@chromium.org>
> Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1101280}

Bug: 1412210
Change-Id: Ie04dbeb6d509ed7623ef2a14681e9d1e969b76af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4224957
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Phillis Tang <phillis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102393}