[history] Fix unchecked chrome.send param in ForeignSessionHandler
The bug describes that a heap overflow can be triggered by using
chrome.send to send an invalid message to the C++ handler.

That was true because the C++ handler did not check the bounds of the
array before using the parameter.

This code is essentially unowned, but I'm fixing it because I'm listed
as a History owner.

Bug: 1408120
Change-Id: Ia8f049ca8bc35bbe63122affcb24b83d7d9cdb62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4226314
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Demetrios Papadopoulos <dpapad@chromium.org>
Auto-Submit: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102408}