Clean up the WinUseBrowserSpellChecker feature.
The WinUseBrowserSpellChecker feature is always enabled in production.
It existed only to disable the browser spell checker in tests. This CL
replaces it with a boolean that can be manipulated in tests.

This eliminates crash reports generated in production when the state
of the WinUseBrowserSpellChecker is accessed before the FeatureList
is initialized.

Bug: 1386223, 1392145
Change-Id: I51d15a826a84fce25c47733acb12b61beea72698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4188827
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102419}