safe_browsing: Handle invalid 0-length blkx plists in a DMG
A plist is required, but if it were empty, the code would erroneously
access the front of a zero-length vector. Switch to using
std::vector::data(), which is safer, and add an explicit length check.

No-Presubmit: True
Fixed: 1413584
Change-Id: I4b60a297cd401369852ba852963fdcfd7a065283
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4228512
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102421}