Add some worked examples for safe math to base unit tests.
It can be confusing to reason about the behavior of the
templated safe math operations, when differently typed
parameters are supplied to each argument.

This CL adds some worked examples as a test for CheckMul to
illustrate the behavior of all the different variants.

BUG=332611

Change-Id: If69e61ef2fdd8957105d96853e19a0b44c67008a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4226303
Reviewed-by: Gabriel Charette <gab@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102455}