Mark sandbox test only code as testonly in build file.
This CL makes a number of improvements/fixes to the Windows
sandbox BUILD.gn file.

1. :common should not depend on :common_test since it's test only
code, so the dependency was reversed. The name :common_test was
also changed to :test_support to conform with naming standards
elsewhere in Chromium.

2. Several targets were marked `testonly` so they can depend on
:test_support.

3. tests/common/controller.h was moved to :common to :test_support
since it's test code and not production code.

4. Dependencies to //sandbox were changed to :sandbox to avoid
having to jump through the public_dep in //sandbox for no reason.

BUG=None

Change-Id: I82f3268fb3e265997914d7511ec12383ddb5a5cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4228792
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102493}