flags_ui: Add comment for platform feature name restriction
According to go/cros-platform-finch, the platform feature requires
starting with a prefix. If this is not followed, there will be no build
error which might waste developer's time when they are not familiar with
it.

This CL adds a comment to the pointer to the document for better
readability. Also add unit test to check name Validation.

BUG=None
TEST=AboutFlagsTest

Change-Id: I23623a85689b8429543b4e7d0e8ecfafede04542
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4222836
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Shengsong Tan <sstan@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1102506}