LacrosTablet: Send FlingScrollEvent for TabScrubber to Lacros
This CL lets HandleTabScrubbing crosapi pass FlingScrollEvent as well,
not only scroll event which was originally filtered by finger count.
This event is necessary to release `immersive_reveal_lock_`.

It could be better to handle this in wayland instead of crosapi in the
long term, but it requires to introduce additional wayland extensions
(allow sending event for active window, send finger count), so for now
we expand crosapi to resolve the bug.

Bug: 1411485
Change-Id: I11cb90ac99db39d1fb38351179dfea19dd05fa08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4216537
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102516}