Fix \ != / presubmit failures on Windows
remap_paths_to_relative uses os.path.relpath which uses the local
machines path separators, which means that we get different results on
Windows from on Linux. These different results cause presubmit failures
on Windows which is good, because it reveals the problem.

This change forces the path separators to be '/' for consistency.

This change also adjusts how a text file is opened for writing to avoid
future compatibility issues, in line with the Python style guide.

This fix was tested on Windows with this command:
  git cl presubmit --force --files testing/merge_scripts/code_coverage/merge_js_lib.py

Bug: b:265968684
Change-Id: Ia9d98b6e95381d3f0f4d0c201d503b7fe9ba1feb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4229413
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102525}