Separate out the common ash/lacros parts of DriveFsNativeMessageHost
This is done to prepare for the lacros version of this message host. The
shared code is split into chrome/browser/chromeos/drivefs where it can
be accessed by both ash and lacros. This also splits out the native
messaging interfaces of the drivefs mojom so they can be accessed by
lacros.

Bug: 266163193
Change-Id: I6811b83dfac6dbefae219675962966502199a2c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4191005
Reviewed-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Austin Tankiang <austinct@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102527}