DLP: Only show banner for file-open dialogs
The DlpRestrictedBanner for file-open dialogs is shown when some of the
files in the visible list are DLP blocked. This is based on the state
we get from the store and that only checks if any of the listed files
are blocked from being opened by the destination, and doesn't consider
the dialog type which caused an empty banner to show in e.g. select
folder types. With this change, we only update the banners for the
correct dialog types.

Fixed: b/267621719
Change-Id: Id6b95dcfd5019fb4de08579dde215dbf477da2fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4223377
Commit-Queue: Aida Zolic <aidazolic@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102618}