[Profiles] Add new values to ProfilePicker::EntryPoint enum
This CL splits the entry point kOnStartup and kNewSessionOnExistingProcess
in two, as these were used from two different places.
The new values are used when the ProfilePicker was not requested, but is
shown because no suitable profile was found.

Bug: 1150326
Change-Id: I22c576f7d1921cd4bbfd1335249ad00980d539e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4227078
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102624}