[Privacy Sandbox] Set focus to blocked section after removing last item
This CL fixes an accessibility bug when blocking or allowing the last
topic/site on the Topics/Fledge settings subpage.

The focus was lost and moved to an empty region after the last item on
the list was removed.

The focus is then set to the blocked topics/sites section row instead in
these cases.

Preview:
https://screencast.googleplex.com/cast/NTI2NzA4ODY2MzkwNDI1Nnw3OTM2ZjI2Zi00Mg

Bug: 1414038, b/267355079
Change-Id: I26fbeec8a8ef7ec60fbe1a2ef95a4e434f6cdeae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4231815
Reviewed-by: Olesia Marukhno <olesiamarukhno@google.com>
Commit-Queue: Gabriel Oliveira <gabolvr@google.com>
Cr-Commit-Position: refs/heads/main@{#1102714}