Fix SW checksum update happens twice when ChecksumUpdateTiming::kAlways
This is a follow-up CL for the parent CL[1] and the discussion in the
other previous CL[2]. Currently the checksum can be updated twice, and
it hits DCHECK when ServiceWorkerCacheWriter runs with the
ChecksumUpdateTiming::kAlways option and there is an update in SW
scripts.

Specifically, that case would go with the following scenario.

1) SW resources are stored, but sha256 checksums in the database are
empty.
2) The update check runs with the ChecksumUpdateTiming::kAlways option to
update the sha256 checksum.
3) ServiceWorkerCacheWriter finds a difference between cached data and
data from the network, then checksum->Update is called twice in
`MaybeWriteData` and `WriteDataToResponseWriter`.
4) ServiceWorkerCacheWriter instance is moved and `GetSha256Checksum` is
called from ServiceWorkerUpdatedScriptLoader.

To solve this issue, this CL adds 1) the if statement to
`WriteDataToResponseWriter` to call checksum->Update only when
`ChecksumUpdateTiming` is `kCacheMismatch`. 2) delete the
ChecksumUpdateTiming DCHECK in `ServiceWorkerUpdatedScriptLoader` since
it can be both enum values. 3) unittest to check this scenario.

[1] crrev.com/c/4209712
[2] https://chromium-review.googlesource.com/c/chromium/src/+/4210872/comment/8d9915da_7b2be08d/

Bug: 1371756
Change-Id: I2ef0fa1a703a5a31690201174b42c6df68317ce5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4227397
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1102720}